Create Tick-Tack-Toe_4_5_final
There was an issue with a recursive method that did not want to loop as I was expecting. Although there should have been a workaround using "return ". I ended up dividing the and making it run in the main method.

Given the inconsistent manner of getting to know how the input is going to be, the code is not ideal a) too many conditions b) too many references and conversion between the input and output.  The code is not eve close to production level.  

https://hyperskill.org/projects/48/stages/258/implement