Create rock_paper_scissors_EX5_progress
Just posting a funky solution, its not working. I have faith I will be able to figure out the fault tomorrow. 

its something to do with .split getting an empty input.

there is the issue of getting the custom game array from a method. I think my logic is a bit flawed and I need to create a global array.

https://hyperskill.org/projects/314/stages/1768/implement