Create Amazing_numbers_ex2
My initial idea of how many classes I need was a bit too large, I might even trim the classes down even further as we progress. The main takeaway is that if you are going to call a method within a method declare it as a variable and get the return value, my Initial idea of calling the method every time I need the value is moronic however I have been doing so. Live and learn.

https://hyperskill.org/projects/184/stages/932/implement